[Jest] Migrate `spec.test.js` (#647)
